Adds Date and Image extensions
Introduces extensions for Date and Image.

Creates a `Date` extension for consistent JSON date formatting
throughout the app.

Adds an `Image` extension to calculate the average color,
allowing for dynamic UI theming and visual effects.

Renames `Extensions.swift` to `UIDevice+Extension.swift` and moves
device-specific UI measurements to it, maintaining code organization.
